# backend/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Testing Nirvana API
servers:
  - url: api
components:
...
  schemas:
...
    Pagination:
      type: object
      properties:
        pages:
          type: number
        total:
          type: number
      required:
        - pages
        - total
...
    School:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        district:
          type: string
        state:
          type: string
        contactEmail:
          type: string
          format: email
        maxTravelDistance:
          type: number
        maxStudentsPerTest:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - district
        - state
        - contactEmail
        - maxTravelDistance
        - maxStudentsPerTest
        - createdAt
        - updatedAt
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dob:
          type: string
          format: date
        gradeLevel:
          type: number
        school:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required:
            - id
            - name
        parent:
          type: object
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
              format: email
          required:
            - id
            - firstName
            - lastName
            - email
        studentCode:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
        confirmationStatus:
          type: string
          enum:
            - CONFIRMED
            - PENDING
            - RESCHEDULED
        createdAt:
          type: string
        updatedAt:
          type: string
        accommodations:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              details: {}
            required:
              - id
              - name
        assignedTeachers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              firstName:
                type: string
              lastName:
                type: string
            required:
              - id
              - firstName
              - lastName
        assignedTests:
          type: array
          items:
            type: string
      required:
        - id
        - firstName
        - lastName
        - dob
        - gradeLevel
        - school
        - parent
        - studentCode
        - status
        - confirmationStatus
        - createdAt
        - updatedAt
...
paths:
...
  /schools:
    get:
      description: Get a list of schools
      parameters:
        - schema:
            type: string
          required: false
          name: name
          in: query
        - schema:
            type: string
          required: false
          name: district
          in: query
        - schema:
            type: string
          required: false
          name: state
          in: query
        - schema:
            type: string
          required: false
          name: contactEmail
          in: query
        - schema:
            type: number
          required: false
          name: maxTravelDistance
          in: query
        - schema:
            type: number
          required: false
          name: maxStudentsPerTest
          in: query
        - schema:
            type: string
            enum:
              - name
              - district
              - state
              - contactEmail
              - maxTravelDistance
              - maxStudentsPerTest
            default: name
          required: false
          name: sortBy
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          required: false
          name: sortOrder
          in: query
        - schema:
            type: number
            minimum: 1
            default: 20
          required: false
          name: perPage
          in: query
        - schema:
            type: number
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
      responses:
        "200":
          description: Object with data and pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/School"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - data
                  - pagination
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    post:
      description: Create a school
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                district:
                  type: string
                state:
                  type: string
                contactEmail:
                  type: string
                  format: email
                maxTravelDistance:
                  type: number
                maxStudentsPerTest:
                  type: number
              required:
                - name
                - district
                - state
                - contactEmail
                - maxTravelDistance
                - maxStudentsPerTest
      responses:
        "204":
          description: The school has been created successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /schools/{id}:
    get:
      description: Get a school
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Object with data containing school information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/School"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    patch:
      description: Update a school
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                district:
                  type: string
                state:
                  type: string
                contactEmail:
                  type: string
                  format: email
                maxTravelDistance:
                  type: number
                maxStudentsPerTest:
                  type: number
      responses:
        "204":
          description: The school has been updated successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    delete:
      description: Delete a school
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: The school has been deleted successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /students:
    get:
      description: Get a list of students
      parameters:
        - schema:
            type: string
          required: false
          name: name
          in: query
        - schema:
            type:
              - string
              - "null"
          required: false
          name: dob
          in: query
        - schema:
            type:
              - number
              - "null"
          required: false
          name: gradeLevel
          in: query
        - schema:
            type: string
          required: false
          name: school
          in: query
        - schema:
            type: string
          required: false
          name: parent
          in: query
        - schema:
            type: string
          required: false
          name: studentCode
          in: query
        - schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
          required: false
          name: status
          in: query
        - schema:
            type: string
            enum:
              - CONFIRMED
              - PENDING
              - RESCHEDULED
          required: false
          name: confirmationStatus
          in: query
        - schema:
            type: string
          required: false
          name: teacherId
          in: query
        - schema:
            type: string
            enum:
              - name
              - dob
              - gradeLevel
              - school
              - parent
              - studentCode
              - status
              - confirmationStatus
            default: name
          required: false
          name: sortBy
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          required: false
          name: sortOrder
          in: query
        - schema:
            type: number
            minimum: 1
            default: 20
          required: false
          name: perPage
          in: query
        - schema:
            type: number
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
      responses:
        "200":
          description: Object with data and pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - data
                  - pagination
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    post:
      description: Create a student
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                dob:
                  type:
                    - string
                    - "null"
                gradeLevel:
                  type: number
                  minimum: 1
                  maximum: 12
                schoolId:
                  type: string
                  format: uuid
                parentId:
                  type: string
                  format: uuid
                studentCode:
                  type: string
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
                confirmationStatus:
                  type: string
                  enum:
                    - CONFIRMED
                    - PENDING
                    - RESCHEDULED
                accommodationIds:
                  type: array
                  items:
                    type: string
                teacherIds:
                  type: array
                  items:
                    type: string
                testIds:
                  type: array
                  items:
                    type: string
              required:
                - firstName
                - lastName
                - dob
                - gradeLevel
                - schoolId
                - parentId
                - studentCode
                - status
                - confirmationStatus
      responses:
        "204":
          description: The student has been created successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /students/{id}:
    get:
      description: Get a student
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Object with data containing student information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Student"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    patch:
      description: Update a student
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                dob:
                  type:
                    - string
                    - "null"
                gradeLevel:
                  type: number
                  minimum: 1
                  maximum: 12
                schoolId:
                  type: string
                  format: uuid
                parentId:
                  type: string
                  format: uuid
                studentCode:
                  type: string
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
                confirmationStatus:
                  type: string
                  enum:
                    - CONFIRMED
                    - PENDING
                    - RESCHEDULED
                accommodationIds:
                  type: array
                  items:
                    type: string
                teacherIds:
                  type: array
                  items:
                    type: string
                testIds:
                  type: array
                  items:
                    type: string
      responses:
        "204":
          description: The student has been updated successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    delete:
      description: Delete a student
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: The student has been deleted successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
...